Refactored BookMeeting folder structure
- Changed the folder structure of src/components/BookMeeting
