+++ src/components/BookMeeting/ReportIssue/Popup/FormPartials/InputTextContainer.jsx
@@ -1,14 +1,14 @@
 // MUI Imports
 import { Container } from '@mui/material';
 
-// Used in form for select fields
+// styles
+const containerStyles = { display: 'flex', alignItems: 'center' };
+
+// Used in report issue form for select field
 // Sets the text + icon to be on same level
 const InputTextContainer = ({ children }) => {
   return (
-    <Container
-      disableGutters={true}
-      sx={{ display: 'flex', alignItems: 'center' }}
-    >
+    <Container disableGutters={true} sx={{ ...containerStyles }}>
       {children}
     </Container>
   );
