+++ src/components/BookMeeting/ReportIssue/Popup/Main.jsx
@@ -2,20 +2,22 @@
 import { Dialog, DialogContent } from '@mui/material';
 
 //Local Imports
-import PopupTitle from './PopupTitle';
-import PopupButtons from './PopupButtons';
-import PopupFormFields from './PopupFormFields';
+import PopupTitle from './PopupPartials/Title';
+import PopupButtons from './PopupPartials/Buttons';
+import PopupFormFields from './FormUI/Main';
 
 // This is the main component, creates a popup with a form
 // for reporting issues with the app (or smth. else)
 const ReportIssuePopup = ({ open, handleClose }) => {
   return (
     <Dialog open={open} onClose={handleClose}>
+      {/* Popup main body */}
       <DialogContent>
         <PopupTitle />
         <PopupFormFields />
       </DialogContent>
 
+      {/* Popup buttons/ footer */}
       <PopupButtons handleClose={handleClose} />
     </Dialog>
   );
