+++ src/components/BookMeeting/ReportIssue/Popup/PopupPartials/Title.jsx
@@ -2,19 +2,19 @@
 import { DialogTitle, Typography } from '@mui/material';
 import GppGoodOutlinedIcon from '@mui/icons-material/GppGoodOutlined';
 
+const iconStyles = {
+  background: '#3d50af',
+  color: 'white',
+  borderRadius: '50%',
+  padding: '5px',
+  marginRight: '10px',
+};
+
 // Header of the popup card
 const PopupTitle = () => {
   return (
     <DialogTitle sx={{ display: 'flex', justifyContent: 'center' }}>
-      <GppGoodOutlinedIcon
-        sx={{
-          background: '#3d50af',
-          color: 'white',
-          borderRadius: '50%',
-          padding: '5px',
-          marginRight: '10px',
-        }}
-      />
+      <GppGoodOutlinedIcon sx={{ ...iconStyles }} />
       <Typography component="span">Report an issue</Typography>
     </DialogTitle>
   );
