+++ src/components/BookMeeting/Title/Main.jsx
@@ -1,0 +1,16 @@
+import TitleText from './TitleText';
+
+const Title = ({
+  location = 'Orange {kITchen}',
+  meetingRoom = 'Agora',
+  meetinRoomStatus = true,
+}) => {
+  return (
+    <>
+      <TitleText>{`${location} - ${meetingRoom}`}</TitleText>
+      <TitleText>{meetinRoomStatus ? `Free until 10:50` : `Busy`}</TitleText>
+    </>
+  );
+};
+
+export default Title;
